DHCP hook push to a Cloudflare Worker instead of local processing.
Offloads the issue of operating a KV store from ourselves to Cloudflare. For five bucks a month, fuck it.
The KV store is cool and I might use it for other things that is not just a DHCP daemon so the whole quota will not be wasted on just this.